backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
onnxruntime>=1.18.0
Pillow>=10.0.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
